Implemented and tested unarchive()
First unarchive() was going to use Archive::Tar and Archive::Zip, but
Archive::Zip is needlessly complicated for my needs, and recommended the
excellent Archive::Extract, which I now use exclusively.

Archive::Extract does not depend on Archive::Zip, but it will use it if it is
installed, which is needed for Windows support.