Implemented and tested fetchware() and override()!
fetchware() and override() are done! Now I can move on to bin/fetchware. This
release includes a bug I need to fix where executing fetchware() will cause the
program to be installed this needs to only be done when bin/fetchware asks for
it to be installed to support the cool package manager stuff.

This commit also fixes a bug in lookup() and its associated test file where I
commented out stuff instead of fixing the bug in the first place, but that bug
is now fixed.

Implementing override() was a breeze thanks to Sub::Override being surprisingly
awesome and lexical, so I won't mess up other stuff, because things will only be
overriden whil override() is being executed, and then after override() returns
all of the overriden subs return to normal! It's awesome and powerful!

Why bother with OO and subclasses and roles and and Moose, when clever
overrideing of subroutines is easier, faster, and better.