Replaced direct access of %CONFIG with a config() subroutine.
Globals are evil! Must be destroyed! :)

I destroyed the last remnants of the misguided %FW variable. It was a
leftover from fetchware's previous design, and now its gone forever!

It was replaced by config(), and config() helper functions such as
config_replace(), config_delete(), __clear_CONFIG(), and debug_CONFIG().

All of those functions are within a bare block that contains %CONFIG to
limit the scope to only those subroutines.