Finished implementing and tested bin/fetchware's utility subroutines.
Had to determine where to store the fetchware database. The code uses
File::HomeDir, but it requires testing on Windows, Mac OSX, and everything
else to be sure it works properly on those systems. Especially Mac OSX: How
do I determine root on OSX? And where is the OSX location for
/var/log/fetchware?

Next, I'm going to test new(), cmd_install(), and so on.