Added msg() & vmsg() calls, and replaced systemx() with run_prog().
Removed use of systemx(), and replaced it with fetchware's own run_prog().
run_prog() supports -q for silencing even the output of external commands
that are run. It just uses a piped open() to read their output from stdout,
and then it just ignores the output.

These functions could support logging functionality later on, but for now
the only logging is using the shell to redirect stdout and/or stderr if
you want to.

Next, I'm going to design, implement, and test fetchware's replacement
override system that isn't going to actually override anything like the
crazy old one did with Sub::Override, which is mostly meant for testing
not abused to create a plugin system.

Then, fetchware will be finished! Except for all of the remaining bugs that
still need to be squashed.