Implemented and tested override functionality.
Override functionality has been implemented in a different way than I thought I
would. I was stuck copying and pasting a 15 or so line chunk of to all of the
App::Fetchware API subroutines. This chunk of code was annoyingly copy and
pasted, because it needed to use caller(), but from the prespective of the
particular API function. It also needed to return for various reasons, but also
to sometimes just execute everything anyway. Because of all of these reasons it
seemed best to just copy and paste it everywhere: if only Perl had macros. CPAN
probably does, but that would be a pretty silly module requirement. And yes I
already had to add code to the top thing in each and every API function, which
was a real pain, but vim came to the rescue, and it was pretty simple, and not
worth fixing.

The App::Fetchware API subroutines, start(), lookup(), download(), verify(),
unarchive(), build(), install(), uninstall(), and end(), all support being
overridden in Fetchwarefiles using the simple syntax:
start sub {
    Your overridden sub here
};

They all can also be overridden easily using an alternate App::Fetchware import
in your Fetchwarefile:
use App::Fetchware::DirectorySync;
...

This is what I'm going to test next. Then fetchware is pretty much finished
aside from getting dzil working again, and fixing all of the bugs.