skip_all_unless_release_testing() now tests all of FETCHWARE_* env vars.
I considered using /xt for FETCHWARE_RELEASE_TESTING, but deceided that it's
best to keep them integrated with the regular, user tests in /t. This means that
skip_all_unless_release_testing() is still very important to keep software from
being installed on your system in order to test fetchware's ability to do so.