Added tests to support HTMLPageSync's keep_destination_directory.
I previously implemented keep_destination_directory, but I didn't add tests for
it. This commit addes a subtest to test it, and refactores HTMLPageSync.t quite
a bit.

The biggest change was fixing a massive bug that prevented HTMLPageSync from
actually downloading anything. Instead, it was returning undef, and nothing
freaked out about getting undef. I had to fix that bug, and now HTMLPageSync.t
actually downloads stuff to a tempdir that is then copied to
destination_directory where it is later deleted, because destination_directory
is also a temporary directory.