Changed drop_privs() to support communication between child and parent.
Added a pipe to allow the child to communication variables or whatever it wants
to back to the parent. This is needed to make cmd_upgrade() work properly. And
is better than some stupid workaround where cmd_upgrade() manually writes to a
file, and then the parent reads the file.

If I'm going to impose drop_privs() on everyone, then it should be as flexible
as it needs to be.

Next, I'll finish adding drop_privs() support to bin/fetchware, test it, and
then finish the rest of the release blocking bugs.