Finished drop_privs() impl and testing.
drop_privs() has been implemented and tested. It was a real pain. There were
"dependencies" everywhere. Such as the order things must be executed, some
things needing to be shared back to the child, while others must be done before
in the parent. It was a debugging nightmare.

I also implemented cmd_clean() for fetchware clean functionality. It even
involved adding a fetchware semaphore file to each tempdir that is created so
fetchware clean knows if it can delete the file or not.

I also implemented a %SIG handler, but it needs some more work.