Added PATH param support to download_{file, dirlist}.
lookup() is not going to return a path that you then append to mirrors to
determine the download URL. So, I had to modify download_dirlist() and
download_file() to take such a PATH as a parameter.

Next, I'll refactor lookup() and download() to use a path instead of a full url.
Then, I'll refactor verify() to do the same.