Implemented and tested lookup_url, mirror, download_path, and verify() changes.
Now, verify() needs to import keys using a URL based on lookup_url, which should
be the author's server not a mirror, so you can download digests and digital
signatures that the author uploaded instead of ones potentially altered by a
mirror site.

download() also changed to support lookup() now returning a path instead of a
URL, and the new requirement of a lookup_url and at least one mirror, and a huge
recommendation for a gpg_keys_url as well.