Finished parse_fetchwarefile()'s changes in check_lookup_config().
The "less generic" config file checks that I wanted to make I added to
check_lookup_config() instead of parse_fetchwarefile(), because it is more
generic, and only App::Fetchware uses it, so other fetchware extensions are free
to use whatever syntax checking they would like.

Next, I'm going to add support for the lookup() and verify() changes and the
changes to the required config options to cmd_new() in bin/fetchware.