Fixed fetchware help output and dropped adding a 'EXIT' sighandler.
Added look, list, clean, and new to the man page style "SYNOPSIS" at the top of
the help display. Deleted the annoying BUGALERT comments, and rearanged the
description of the commands to match up with the list in the "SYNOPSIS".

I also deleted my use of creating a EXIT sighanlder, because there isn't a
signal named that, so I shouldn't even create one.