Removed newline restriction from drop_privs() pipe helper subroutines.
cmd_new() also uses drop_privs(), and it writes the Fetchwarefile it
generates through the pipe to the parent, and this Fetchwarefile obviously
has newlines in it. The old helper functions used newline to split up the
list of variables that was written across the pipe. The new implementation
in this commit replaces the newline, with a randomly generated at
compiletime magic number, and this is used to split up the stream from the
pipe into variables.

The old subroutines pipe_{read,write}_newline() were also renamed to
{read,write_dropprivs_pipe(), because they no loner use newlines.