Fixed previous bin/fetchware new bugs.
The bugs mentioned 2 commits ago have been fixed. Except for the cool cmd_new()
test using Test::Expect. I just could not get it to work properly. I tried a
bagillion and one values for the prompt, but it just won't work right. It's
weird the same prompt that works great in the other Test::Expect tests does not
work in cmd_new(). I believe the problem is that the one prompt prints a
gigantic Term::UI print_me that might be bigger than what Test::Expect uses to
find the next prompt. But I don't know, so the code has been suitabley marked
BROKEN.

I also fixed a bug in the %SIG handler code. Now, the sig handler is only run in
the parent process when a CTRL-C (or CTRL-Z on Windows) is received. Before it
was run in both resulting in weird errors on the child, because the file handles
used vanish under the childs eyes. This was fixed by only doing it in the
parent. I also had to fix my code to ensure it also works when drop_privs() is
not used.

Next up I'll create more example Fetchwarefile's and add an EXAMPLES section to
App::Fetchware listing these examples.