Added an example Fetchwarefile for PHP.
This was a real pain, because PHP uses a Web application instead of a simple
fairly standard mirror system. There is no simple auto index page for fetchware
to use, so a custom lookup handler is needs as well as a download and verify
handlers as well.

I also deleted the junk nginx.Fetchwarefile file. It's been copied and pasted
into App::Fetchware so I don't need it anymore.