MariaDB Fetchwarefile that uses version control instead of Web scraping.
This cool Fetchwarefile uses MariaDB's version control system bzr for see if new
versions are available. Unfortunately, I can't get it to work. Either because of
a bug I filed (https://bugs.launchpad.net/bzr/+bug/1232992), or perhaps because
the dirstate helper thingy didn't compile right on my Linux setup, so I had to
skip it when I installed bzr, which may have led to this annoying bug preventing
me from using bzr correctly with MariaDB.

I'm commiting this untested Fetchwarefile, because maybe I'll be able to get it
to work better in the future, and it's pretty cool.