Add another php example Fetchwarefile that uses git VCS instead of mirrors.
I was going to use MariaDB's VCS for a cool VCS example, but unfortunately
it uses bzr, which is written in python, and is shockingly slower than git
is. I know bzr is in python, but taking 1.2gigs of ram and 45 minutes for what
is supposed to be a "lightweight" checkout is appallingly crummy.

Anyway, so I created a VCS example using PHP, which uses the much much
faster git VCS tool.

Doing so, I fixed a bug in lookup_by_*() to make them more generic, so
Fetchware extension and hook writers can actually make decent use of them.

In Util, I added another stupid chunk of code to clear bizarre taint errors.
Fetchware does not run in taint mode, and does not support it currently, so
how do taint errors pop up. WTF!?! Instead, if finding and fixing this bug
in perl, or in one of the CPAN modules I uses that messes with perl's taint
mode to support actually running in taint mode, I've just added a stupid
regex to clear any possible taint errors.

Fetchware does not run in taint mode, because it would server no real
purpose. Fetchware's design directly string eval's each Fetchwarefile, so
running taint mode is not going to work. What regex is going to verify that
the Fetchwarefile does not have any security problems. None is, so taint
mode is incompatible with fetchware's design.

In Util, a weird bug is fixed to avoid drop_privs() creating an unnecessary
tempdir when start() is overridden.

Next, I'll add the Postgres Fetchwarefile example.