Deleted overly complex but cool BUGALERT regarding lookup_by_versionstring().
It would have required weird code duplication to work with by an arrayref
or an array. Perhaps I could have implemented it as a separate subroutine,
but then there would still be code duplication between them.

So, I ditched this idea, because it's too complicated to implement and
maintaining it is going to be insane.