Fixed bug in drop_privs() SkipTempDirCreation code.
The previous unless was broken, because it would look for fetchware at
the beginning of the line, which wouldn't work with the regex used.

It's easier to just use the basename(), then it is to modify the regex to
support multiple things separated with slashes, but only look in the last
one.