Added PostgreSQL example Fetchwarefile to list of examples.
It is a good example, and shows how inflexible fetchware's default lookup()
subroutine is. But it also shows how easy overriding it is.

Next, I'm going to add a test that will get its own FRT, because I don't
want it to run every time i run dzil test. It is going to actually run
fetchware install on all of the example Fetchwarefiles that I've added in
recent commits! It should be a great automated test of Fetchware's actual
behavior on real usable customized Fetchwarefiles.