Fixed POD error.
I don't know who thought requiring POD having blank lines around it was
acceptable, but it's not. I think its fixed in newer parsers, but the one
I'm using right now, and sorta annoyingly stuck with thanks to "backwards
compatibility," has this stupid bug. So, I'm stuck with this nuisance.