Polished HTMLPageSync for release.
However, a pretty serious bug exists in it breaking how it should work. It needs
to only download "new" links since last screen scrap, and currently it does not.

Crap! This might be a *big problem* for upgrade(), because upgrade() uses a
simple eq to determine if a new version is available. Should I mess with this
just to add support for this? Or just have lookup() do it *"manually"*?

I'll probably go the "manually" route for now, because it shows how flexible
fetchware is without some stupid annoying change. But adding a upgrade() API
subroutine in the future might be a good idea.