Fixed last commit regarding using tmpdir() as cmd_clean()'s default.
The last commit would cause cmd_clean() with no arguments to do nothing, which
is not what it is supposed to do. When no arguments are specified it should
clean up the default temp_dir, which is the system default tmpdir() as returned
by File::Spec. This commit fixes this.