Fixed t/bin-fetchware-util.t bug where a test was too specific to my computer.
I used an eq test instead of a like test, which messed up the test, because I
tested for a specific value that only exists on my computer, and will obviously
differ on other computers.

This is also the first in a series of commits to fix bugs discovered by the
awesome CPAN Testers folks! Thanks for the test report!