Fixed optional Test::Expect bug in t/bin-fetchware-new.t
Test::Expect is "optional" for testing. That means that only if its "already"
installed will it be used. Perl does have a cool TestRequires mechanism, but
expect is wonky on Windows, and I don't want to worry about that, so I only want
to use Test::Expect if it's already installed. I do not want it installed for
testing like the whole TestRequires thing does.

Fix for test report:
http://www.cpantesters.org/cpan/report/48b619bc-5235-11e3-88f0-885982088f79