Fixed "masks earlier declaration" bug in t/bin-fetchware-upgrade.t.
The copy and paste coding style I always seem to fall into when creating
Test::More style test suites caused 2 variables to have the same name. Renamed
the second one. Problem solved.

I've considered creating some sort of macro system for creating easily
repeatable testing code similar to copy and pasting, but without the actually
coping and pasting. But would such a system be stuck using a source filter :( ?
Or use subroutines that are called with args, and create a fancy easy to use
function factory to make them all??? Would it even be worth it. A macro system
would also fail the grep test, which isn't always such a big deal.

Fix from test report:
http://www.cpantesters.org/cpan/report/48b619bc-5235-11e3-88f0-885982088f79