Deleted # PODNAME line to attempt to fix CPAN indexing.
I used the info at the blog article below to make bin/fetchware indexable on
CPAN:

http://blogs.perl.org/users/polettix/2011/11/distzilla-podweaver-and-bin.html

But it recommends using *either* a package or a # PODNAME, and I'm bizarrely
using both, which is perhaps what is screwing up PAUSE's attempts at indexing
me.

I think the package fetchware; line was added later on after # PODNAME was
added, which is now this happened. I didn't add both at the same time.

By the way, the indexing that's broken is this page does not exist:

https://metacpan.org/module/fetchware

But it should, because I do have a package fetchware; line in bin/fetchware. So,
why isn't it working? I believe it's due to not being correctly indexed by
PAUSE, which basically seems to make the file invisible. Before attempting to
email CPAN's maintainers it makes more sense to just try out this simple
solution first.