Fixed bin/fetchware not being indexed by adding missing shebang line.
Maybe that will do the trick. I though it was dzil that adds this, but dzil does
not touch it. Instead, Module::Build or ExtUtils::MakeMaker does this for you.
That means that I need it to have a shebang line in the script, and the build
scripts will just update it to $^X during install.