Fixed sort block returning undef bug.
When I fixed the unequal number of digits bug in lookup_by_versionstring(), I
completely changed the algorithm. This algorthm assumed that perl was okay with
sort blocks returning undef, my Perl setup tested fine, but according to CPAN
Testers the warning was expaned in a later version of Perl, and the warning is
even a fatal error on some Perl setups the CPAN Testers tested fetchware on.

So, this patches fixes that by discarding any files from @versionstrings that
have an empty versionstring, so sort no longer sees any undefs to warn or die
over.