Fixed github issue #1: "uninitialized" error message in  fetchware new.
When gutting fetchware of all cool given/whens I switched around Aristotle's
cool forking idiom
(http://blogs.perl.org/users/aristotle/2012/10/concise-fork-idiom.html) from the
one that uses given/when to the one that does not. I changed the "if not defined
$kid" part to something else, so I could use a fetchware style heredoc long error
message, but I commented out the right code, and the test code to replace it was
wrong. It's functionally the same, but it spits out annoying error messages,
because I use $kidpid before checking its definedness. Now, I just stick with
the more correct, but kinda wonky looking "if not defined $kidpid".