Fixed github#2: new param not asking user name.
Like the previous bug it was also caused by my rushed removal of my use of
given/when. I completely messed up the arguments to cmd_install() and
cmd_upgrade_all() causing the weird bugs the poster mentioned.

Annoyingly, I do not yet have a test suite for fetchware's command line
interface or even the run() subroutine that bin/fetchware uses to be a
modulino. I'll have to add one in the future.