Fixed stupid perl HEREDOC parsing bug.
This might actually be a bug in Perl's parser! I don't really think it's worth
reporting to p5p, and chances are it's actually a bug in my code.

The diff was left in, so that you can see what the stupid bug was. Notice the --
lines that's what the code was, and the + lines are the code now. The fix was
whitespace between the parent and the "prompt" hash key.