Fixed copy_fpkg_to_fpkg_database() test failing to delete a test fpkg.
The left over was named App-Fetchware when I test it in my git repo, but in a
CPAN download the file is named App-Fetchware-1.008 or whatever version of
fetchware your testing.

Fetchware creates lots of files that should be deleted after the test is run,
managing these manually like malloc()ing and free()ing memory in C is stupid.
Just as it works in C, it works ok in fetchware. But a garbage collector would
be nice: Perhaps an object that stringifies to the path of the file it
represents, and on instantiation it adds that object to a list of stuff to
delete. And then an END {} block comes along and deletes it all for me.