Fixed "uninitialized variable" warnings when fetchware is called by itself.
When just fetchware is run the help command is run, but inside run(), the if's
that determine what cmd_*() subroutine to run spit out annoying "use of
uninitialized variable..." warnings, because @ARGV is undef, and therefore so is
$command.

I also deleted the weird space before  the help message is printed. Most *nix
commands don't do this so fetchware shouldn't either.