Added extra testing for equivilent version strings to lookup_by_versionstring().
lookup_by_versionstring() uses the numbers in program versions, such as
(2,2,4) in httpd-2.2.4.tar.bz2, but sometimes there will be files such as
CHANGES_2_2, which is shorter than 2.2.4. Because it's shorter it is "smaller,"
and therefore, can be short circuited.

This commit now does this to avoid annoying "use of uninitialized variable"
warnings. Also, those warnings may have actually been a bug, because they're
undef, and undef is 0 when used as a number, which may actually mess up how they
should be sorted.

NOTE: This commit changes the previous behavior of 2.0.0 and 2.0 being equal.
Now 2.0.0 is "bigger" than 2.0. I don't think this change is a big deal, or
would surprise anyone. If the previous behavior is wanted instead, I could
change the -1 and 1's to 0's.