Implemented and tested new upgrade() API subroutine.
The way fetchware determines if a new version is available is brittle and lame,
and might not work well in all cases, so I always wanted to give the user a way
to change how fetchware determines if the latest version is newer than the
currenlty installed version.

Now you can just hook it, and change it if your software package does not have
what fetchware calls "versionstrings" in it.

Next, I'll implement and test check_syntax() and new(), and also fix some broken
CPANTesters reports too while I'm at it.