Implemented and tested extension_name().
extension_name() is a "singleton". But not an actual Singleton object in java,
so I can still use global variables. Instead, it's just a subroutine that when
called the first time takes its first scalar, and saves and returns it back to
the caller. Then whenever this subroutine is called again, the same value is
returned back to the caller. This is implemented using a Perl 5.10 state
variable to avoid extra stupid parens to use a closure to fake the state
variable.

extension_name() is used to customize new()'s API subroutine's question text
that is printed to the user and error messages that the user might see, so that
the user sees 'App::FetchwareX::ExtensionName' instead of just the perhaps
confusing App::Fetchware.