Modified fetcwhare to use extension_name().
Originally I was going to use extension_name() everywhere it said 'Fetchware',
but I changed my mind. The use of 'Fetchware' does sorta mean
App::FetchwareX::ExtensionName already. So why change it? The downside is
App::FetchwareX::ExtensionName is gigantically long compared to just
'Fetchware', so it would screw up the the lines of the messages being less than
80 characters.

Instead, it's just used for App::Fetchware in a few error messages. I'm still
going to keep using it. It's self documenting, and it does make the few error
messages that use less confusing. So I think it's worth it, so I'm keeping it.