Refactored get_{mirrors,verification,filter_option}.
I removed them calling append_options_to_fetchwarefile(), and instead have them
returning hashrefs, and then the caller or the soon to be written new_helper()
will do it for you.

I also renamed them all "get" instead of "add" or "determine". This is much
clearer, because now they return the filter or mirrors or whatever instead of
doing the appending to the Fetchwarefile for you.

This change was done to make them more flexible. Now extension authors can use
these subroutines, and instead of just appending the result to their
Fetchwarefile, they can do further analysis or use some configuration options,
but ignore others, or perhaps use the sames ones, but rename them.

All of these options may or may not be supported by new_helper(), which will
remove the tempation to just copy and paste all of the stupid repetitive calls
to append_options_to_fetchwarefile().