Implemented non-OO App:Fetchware::Fetchwarefile. Will go OO.
I've been struggling with having Fetchwarefile go OO or not. Unlike Config,
where because of the insane implementation of Fetchware's configuration
subroutines, and how they are implemented, Config is stuck with its crazy global
implementation. Furthermore, changing it now, would at best result in a per
module configuration scope, which is not much of an improvement over per
Fetchwarefile Config scope, which is what it's current configuration has. Also,
if I do change Config, I'll have to change much of the test suite to match the
new OO calling style, or create a backwards compatibility layer, or something
else horrible. Instead, Config stays non-OO, and very global.

Fetchwarefile, however, is going OO. It just makes sense. The only reason not to
go OO, is to leave it non-OO on purpose, because OO sucks. OO does suck, but it
can be great, or exactly what you want, and it's exactly what people expect.
App::Fetchware is staying its crazy old self, but App::Fetchware::Fetchwarefile
is going OO.

fetchware_config_options() and fetchware_header() vanish to be replaced by
new(), which is obviously a constructor. It takes all of the info that the other
regular subroutines used to take, but does it in one go. Fetchwarefile objects
will be immutable without any setters, because I don't need them, and if you
want to change them, just create a new Fetchwarefile object.

Then, fetchwarefile() will be replaced by generate(), and then they'll be a
config_option() setter, for adding options to you Fetchwarefile object.

And a __debug() to print debugging info out.

There will be no __set_fetchwarefile(), because now it's lexical instead of
global, so you can just create another Fetchwarefile object if you need to.

The next commit will change Fetchwarefile into OO. I just wanted to commit its
old non-oo interface in case I want to switch back to it for some reason.