Fixed sill Win32 file extension incompatibility.
I was foolishly using the static string 'perl' inside a subtest for
run_star_commands(). At first I though that maybe perl was not in the smoker's
PATH, but the output printed both the location of where perl was in the $^X
variable, and it also printed the PATH. But perl was in the smoker's PATH. Then
I figured since its called perl.exe instead of just perl on Windows that maybe
that was the problem especially since run_star_commands() and run_prog() both
use system() directly instead of something that uses that platforms version of
the shell. This may be the problem, because the shell is probably what allows
you to call commands on Windows without always using the .exe extension, and
using system uses whatever the Windows' equivelent version of directly executing
a program on Windows is. So, without the shell to add the .exe, my call of perl
fails.

I fixed this by simply using $^X instead of perl.

Also, this bug came to me in my Inbox from the fine folks at CPAN Testers. The
report is available at:
http://www.cpantesters.org/cpan/report/a6d26844-6c07-1014-9005-f13e6d8a8eb2

Note, this commit only fixes this simple bug. No investigation into the other
test failures has taken place yet.