Fixed flaw in previous check_syntax() and parse_fetchwarefile() impl.
parse_fetchwarefile() had a hard coded number of api subroutines in it set
to 9 in it, but now there are 11, and I'm about to add a 12. So, I updated
to use 'scalar keys %api_subs' instead of a hard coded value. Now, I only
have to add new API subs to %api_subs without worrying about the stupid
easily outdatable hard coded value.

There are also a few documentation fixes to check_syntax(). It's previous
documentation was total crap. I also removed the useless, and extremely
confusing $fetchwarefile argument from check_syntax(). It can no longer
check the actual $fetchwarefile text, but it can still test it via
config(), which was the whole point to begin with.