Refactored verify()'s crazy if to remove annoying 'uninitialized' warning.
Verify used [... eq undef] to test if config('verify_method') had been
defined to something, and because config('verify_method') is not inside a
defined() call it spits out an annoying "use of uninitialized value'
warning. These are annoying, and can freak out any users not used to how
often and annoying these warnings are when coding in Perl. The worst part
about them is occasionally one actually is a bug, so it's not like you can
actually turn them off even if you want to, because eventually doing so
will mask a real bug that will require vials and vials of antivenom to
avoid killing you.