Updated ExportAPI to support the recently added API subroutines.
Since ExportAPI was last modified upgrade() was added! Also, check_syntax was
added, and something else was added too. I also fixed the same bug that
check_syntax() or parse_fetchwarefile(), I forget which had, where the number of
API subroutines was a set ingteger constant. Now it is a proper dynamically
genreted expression based on %api_subs. If more API subs are added, I do have to
add them to %api_subs, but I won't have to update that stupid integer constant
anymore, because now that value is generated with keys in scalar context.

new() and new_install() were not added, but I forgot to added exceptions to
throw when you attempt to import them, because unlike the rest of the API
subroutines they are not "inheritable". They need to be very specific for each
extension, so reusing them is not allowed, but you can reuse their components
that you can load with :OVERRIDE_NEW and :OVERRIDE_NEW_INSTALL.