Deleted out of date commet regarding insane old App::Fetchware hook feature.
In one of the first implementations for App::Fetchware, hook did not exist, and
instead the API subroutines themselves could be called with a coderef argument
inside a Fetchwarefile to hook themselves.

This was a cool feature, but any sort of implementation for it is wonky, and it
required a paragraph explanation before a crazy "package fetchware;" line inside
HTMLDirSync's test suite. This retarded work around would even be required
inside user extensions, which is too stupid and copy and paste for my tastes.

This crazy feature no longer exists, because that past implementation was
removed for that, and other reasons such as being pretty much unimplementable
without stupid copy and paste garbage hence hook exists instead.