Removed KEEP limiatations governing new() and new_install() API subroutines.
2 commits ago, I added error checking to ExportAPI to make it complain about
trying to KEEP or "inherit" App::Fetchware's new() or new_install(). I've since
decided not to bother with any such limitations mostly because, virtually any
App::Fetchware extension is going to want to inherit or KEEP new_install(),
because new_install() asks if the user wants to install the Fetchwarefile, and
does so if they say yes. Why would you not want to do that? You wouldn't, so of
course everyone including HTMLDirSync is going to inherit new_install().

The problem is inheriting new() is not really going to be beneficial at all
ever, and you'll definately end up with App::Fetchware Fetchwarefiles instead of
whatever your extension is called Fetchwarefiles, but it beats having annoying
limitations with crappy implementations.