Fixed confusing typo in get_lookup_url() Term::UI print_me.
Previously it was asking your to type in the lookup_url to check "if a new
version of fetchware is released." Which is not at all what you need to do.

You need to specify a lookup_url that is for whatever program you want to
use Fetchware to install, and automatically install new versions as they
are released.