Removed $fetchwarefile argument from App::Fetchware check_syntax() API sub.
I originally was going to allow check_syntax() to be able to "see" the
Fetchwarefile text, but this has changed, because who wants to try to use
sucky regexes for even sucky PPI to parse perl? I don't.

Instead, Fetchware's check_syntax() is only for "higher-level",
"application-leve", "Fetchware-level" syntax errors.