Added Perl 5.10.1 as a dependency of Fetchware to dist.ini.
Perl 5.10.1 has always been a dependency of Fetchware. It's always had a use
5.10 line in it, and then I changed to 5.10.1, because 5.10 smart-matching was
rewritten, because the version 5.10 was messed up.

Now, I could proabably switch to 5.8 or even 5.6, if I wanted to, because I only
mandated 5.10, because given/when fucking rule. But they use smart-matching, and
smart-matching along with given/when were made "experimental" in 5.16, which
effectvely makes them unusable now. Which blows. I really liked given/when for
testing a bunch of different tests agains one value in the given, and even that
obvious case apparently still uses "stupid-matching" or
"have-to-look-up-what-exactly-it-does-in-the-docs-matching", which is broken
after further reflection on its design.

I'm not going to bother switching to 5.8 or 5.6, because nobody uses them
anymore. I think Red Hat 6 at least has 5.10. I also use // and say
occasionally, and I think some of my dependencies need at least 5.8, and some
perhaps might even need 5.10. So, not it's not worth any sort of effort at all.