Fixed botched AUTOMATED_TESTING SKIP block.
The if statement to skip testing that relies on the terminal working correctly
when run by automated CPAN Testers smokers was broken. It included a not, but it
was an if statement not an unless. I think originally it might have been written
as an unless, and then it was rewritten as a if. Anyway, it should be fixed now,
and this should make at least most of the CPAN testers'FAILs with PASSes.