Attempted fix of weird upgrade errors on some CPAN Testers reports.
The report lacks details on actual cause, but my best guess is the use of a
File::HomeDir default fetchware_package_path() on a smoker that may not have a
proper home directory. It might be a "server" style home dir instead of a proper
"user" style one Fetchware presumes.