Fixed fetchware look drop_privs() temp dir deletion bug
drop_privs() creates a second level of temp dir so what has dropped privs can
write to disk the file it downloads. This tempdir was setup to be deleted at end
of program using File::Temp's CLEANUP option. I have deleted this option now, so
that when the build failes this directory will still be around so that you can
debug the build failure. Prior to this fix the directory would be deleted in
File::Temp's END block preventing you from figuring out what happened.