Fix bug where no_install would still create and even install a package.
The installed package would only be installed into the user's specified
fetchware_database_path() not actually install it onto their system.

But the bug would create the package and copy that package to the user's
fetchware database, which is obviously a huge bug, because the fetchware
database is only supposed to contain packages that have actually been installed,
and if it has not actually installed them, then its corrupt, which might be a
big deal especially when upgrade-all is run.