Fixed php.Fetchware Fetchwarefile test due to HTML page change.
PHP's site has been untouched basically since its inception. It really never
changes, which is awesome for this bit of code :) The only problem is they
recently did actually change their site's HTML to add some more CSS stuff to
make it look better and less 1998. The bug fix was a simple code change. I also
left in my debugging code commented out, because I'll probably need it in the
future, because this code is quite brittle. I should probably code it to
look_down from the a tag until it finds the md5sum, and that should be smarter.