Check cmd_look() to see if $look_path exists after it runs.
Previously, I fixed a bug in commit 65e1d34a8139e3518cdaaa0bf1f793d7a987dca6,
where cmd_look() is broken, and the look path that is printed to the screen does
not exist for some reason.

I forgot to add a test to check for this now, which is commit adds.

However, this commit does not delete the looked up directory, and File::Temp
does not delete it, because cmd_look() tells start() not to, but I just couldn't
get File::Path's remove_path to work, what a total piece of crap. Worst
interface ever. I'm tempted to rewrite it. It's absolute junk.