Removed weird "No errors..." message from cmd_clean() output.
cmd_clean() had a silly, confusing output line of "No errors encountered" that
was from the cut and paste job of using File::Path::remove_tree()'s weird error
handling system. I expanded that weird line into a vmsg() that would only been
seen if verbose output is turned on. I also turned some error messages into
warn()'s instead of msg()'s.

A bug was also fixed in the counting of the number of directories that have been
deleted. It was a mistaken "=" assignment instead of a "++" increment.