Added test alluded to but skipped in 3718b9f for drop_privs()/cmd_install().
3718b9f changes how drop_privs() works in order to fix a bug that happens in
cmd_install() if the cmd_install() fails to build or install for some reason.
This bug fix had no test confirming it fixed anything.

It was this one bugfix that inspired me to rewrite make_test_dist()'s innards
and switch from positional to named parameters. This commit finishes this
original change.

I was originall going to add a test to drop_privs()'s test suite, but it makes
no sense there, so the test for a change to drop_privs() is here in
cmd_install().