Fixed failed expansion of FETCHWARE_DATABASE_PATH to whole file.
This test file tests fetchware by actually running bin/fetchware with real
command line arguments, and by default what will become the users
FETCHWARE_DATABASE_PATH is used, but this is testing so should it really be
used? At first I was thinking yes, but if something is messed up a fake Apache
or test-dist could be left in their fetchware database, and could confuse
fetchware itself or the user. Therefore, the safest choice is to just use
FETCHWARE_DATABASE_PATH for every test in this test file.

This commit fixes the first attempt at making this change, because the previous
use of FETCHWARE_DATABASE_PATH was left in, which includes deleting
FETCHWARE_DATABASE_PATH preventing the last few commands from being tested
properly.